requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.0.6
django-extra-keywords==1.1.2
django-filter==22.1
django-jasmine==0.5.1
django-jazzmin==2.5.0
django-linguist==0.6.0
django-modeltranslation==0.18.4
django-nested-admin==3.4.0
django-phonenumber-field==6.3.0
django-rest-swagger==2.2.0
django-taggit==3.0.0
django-translation==1.2
django-translator==1.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
falcon==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
jasmin==0.10.11
Jinja2==3.1.2
kombu==5.2.4
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
phonenumbers==8.12.52
Pillow==9.2.0
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-messaging==0.5.13
python-mimeparse==1.6.0
python-monkey-business==1.0.0
pytz==2022.1
redis==3.4.1
requests==2.23.0
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
smpp.pdu3==0.6
smpp.twisted3==0.7
sqlparse==0.4.2
tabulate==0.8.10
treq==22.1.0
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txAMQP3==0.9.3
txredisapi==1.4.7
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
